# feat: use ArcStr for storing strings in Schema