--- a/juniper/src/lib.rs
+++ b/juniper/src/lib.rs
@@ -18,7 +18,7 @@ pub use {async_trait::async_trait, futures, serde, static_assertions as sa};
 #[doc(inline)]
 pub use futures::future::{BoxFuture, LocalBoxFuture};
 
-pub use arcstr::{literal, ArcStr};
+pub use arcstr::{self, ArcStr};
 
 // Depend on juniper_codegen and re-export everything in it.
 // This allows users to just depend on juniper and get the derive