@tyranron 2024-04-08
@audunhalland I'm going to review this in next few days and [release in 0.17.0](https://github.com/graphql-rust/juniper/milestone/7).