@audunhalland 2024-04-08
@tyranron cool, I'm not 100% happy with all parts of this PR, there are surely things that can be iterated on later. Especially I see `DynType` and `AsDynType` as a kind of hacky way to limit the scope and get it over the line.