@tyranron 2024-04-08
@audunhalland 

> edit: also note for the review: I'm not sure I understand the `reflect` module fully, like e.g.:
> 
> ```
> impl<S> reflect::WrappedType<S> for ArcStr {
>     const VALUE: reflect::WrappedValue = 1;
> }
> ```

That should be fine, yes.

`reflect::WrappedType::VALUE` is a tricky way to encode composed GraphQL types in prime numbers, since we're kinda limited with `const` evaluation abilities in Rust for now. `1` means a primitive, non-composite type, which `ArcStr` represents, the same way as `String` and similar.