@audunhalland 2024-04-11
@tyranron an idea: Could we somehow make `From<ArcStr>` a supertrait of `ScalarValue`? Then one could make a custom scalar value that doesn't need to clone strings that are already `ArcStr`. Think introspection queries, etc.

(Ideally `From<&ArcStr>` (which doesn't require cloning up front), but that's not possible without making `ScalarValue` lifetime generic. Another idea is to add a method `fn from_arcstr(&ArcStr) -> Self` to `trait ScalarValue` itself.)