@audunhalland 2024-06-17
@tyranron I see there are some compile errors on the bikeshedded changes now. Should I help out with trying to resolve those?