@audunhalland 2025-03-04
This is now one year old, are there any plans for the next version (0.17)?