@tyranron 2025-03-05
@audunhalland yes, this is a blocker for releasing 0.17. Unfortunately, haven't been able to jump on it to finish recently. I'll try to cut some time for it in next weeks.