@tyranron 2025-05-09
@audunhalland I've finally got my hands back to `juniper`, so something like next week this would be finished and released. Definitely would require more than 2 days of effort from me to finish this, but I'm motivated.