@audunhalland 2025-05-14
Nice! I already mentioned further up I too consider `DynType` a hack that temporarily limits the scope of the change.